Move command handling to command-handler, move commands.ts to commands/index.ts
