+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.ts",
   "scripts": {
     "start": "ts-node index.ts",
-    "update": "ts-node src/commands.ts"
+    "update": "ts-node src/commands/index.ts"
   },
   "author": "Lim Yao Jie",
   "license": "MIT",
